requirements.txt
python-telegram-bot
wget
mutagen
vk_api
eventlet==0.22.0
urllib3
bs4